src/tools/templates/main.py
"""Main tool implementation module.

This module contains the primary implementation of the tool's functionality.
Replace this template with your tool's specific implementation.
"""

from typing import Any, Dict, Optional

from .base import BaseToolTemplate


class ToolTemplate(BaseToolTemplate):
    """Main tool implementation class.
    
    This class implements the core functionality of your tool.
    Customize this implementation according to your tool's needs.
    
    Attributes:
        name (str): The name of the tool
        description (str): A brief description of what the tool does
        config (Dict[str, Any]): Configuration parameters for the tool
    """

    def __init__(
        self,
        name: str = "ToolTemplate",
        description: str = "Template for creating new tools",
        config: Optional[Dict[str, Any]] = None
    ):
        """Initialize the tool.
        
        Args:
            name: The name of the tool
            description: A brief description of what the tool does
            config: Optional configuration parameters
        """
        super().__init__(name, description, config)
        self._initialized = False

    def initialize(self) -> bool:
        """Initialize the tool.
        
        Implement your tool's initialization logic here.
        
        Returns:
            bool: True if initialization was successful, False otherwise
        """
        try:
            # Add your initialization logic here
            self._initialized = True
            return True
        except Exception as e:
            print(f"Error initializing {self.name}: {str(e)}")
            return False

    def execute(self, *args: Any, **kwargs: Any) -> Any:
        """Execute the tool's main functionality.
        
        Implement your tool's main logic here.
        
        Args:
            *args: Positional arguments for the tool
            **kwargs: Keyword arguments for the tool
            
        Returns:
            Any: The result of the tool execution
            
        Raises:
            RuntimeError: If the tool is not initialized
        """
        if not self._initialized:
            raise RuntimeError(f"{self.name} must be initialized before execution")

        try:
            # Add your main execution logic here
            result = None  # Replace with actual implementation
            return result
        except Exception as e:
            print(f"Error executing {self.name}: {str(e)}")
            raise

    def cleanup(self) -> None:
        """Clean up any resources used by the tool.
        
        Implement your cleanup logic here.
        """
        try:
            # Add your cleanup logic here
            self._initialized = False
        except Exception as e:
            print(f"Error during cleanup of {self.name}: {str(e)}")

    def validate_config(self) -> bool:
        """Validate the tool's configuration.
        
        Implement your configuration validation logic here.
        
        Returns:
            bool: True if configuration is valid, False otherwise
        """
        try:
            # Add your config validation logic here
            return True
        except Exception as e:
            print(f"Error validating config for {self.name}: {str(e)}")
            return False
