src/tools/templates/tests/test_main.py
"""Test cases for the main module."""

import logging
import pytest
from typing import Any, Dict

from ..main import ToolTemplate
from ..utils.logging import setup_logger

# Set up logger for this test module
logger = logging.getLogger(__name__)


class TestToolImplementation:
    """Test cases for ToolTemplate implementation."""

    @pytest.fixture(autouse=True)
    def setup(self):
        """Set up test cases."""
        self.tool = ToolTemplate()
        logger.info(f"Created tool instance: {self.tool}")

    def test_initialization(self):
        """Test tool initialization."""
        logger.info("Testing tool initialization")
        assert not self.tool._initialized
        success = self.tool.initialize()
        assert success
        assert self.tool._initialized

    def test_execution_without_initialization(self):
        """Test execution without initialization."""
        logger.info("Testing execution without initialization")
        with pytest.raises(RuntimeError) as exc_info:
            self.tool.execute()
        assert "must be initialized before execution" in str(exc_info.value)

    def test_execution_with_initialization(self):
        """Test execution with initialization."""
        logger.info("Testing execution with initialization")
        self.tool.initialize()
        try:
            result = self.tool.execute()
            assert result is None  # Update based on your implementation
            logger.info(f"Execution result: {result}")
        except Exception as e:
            logger.error(f"Execution failed: {str(e)}")
            raise

    def test_cleanup(self):
        """Test cleanup functionality."""
        logger.info("Testing cleanup")
        self.tool.initialize()
        assert self.tool._initialized
        self.tool.cleanup()
        assert not self.tool._initialized

    def test_config_validation(self):
        """Test configuration validation."""
        logger.info("Testing config validation")
        assert self.tool.validate_config()

    def test_custom_config(self):
        """Test tool with custom configuration."""
        logger.info("Testing custom configuration")
        config: Dict[str, Any] = {
            "test_param": "test_value",
            "debug_mode": True
        }
        tool = ToolTemplate(config=config)
        assert tool.config == config
        logger.info(f"Tool created with custom config: {config}")


if __name__ == '__main__':
    setup_logger()
    pytest.main()
